+++ NEWS.md
@@ -1,3 +1,18 @@
+# stdmod 0.2.9
+
+- Fixed the issue with `stdmod-package`. (0.2.8.9001)
+- Improved the printout of the `summary()` of `std_selected()`
+  and `std_selected_boot()` outputs.
+  Small numbers are rounded to prevent
+  the use of scientific notation, and
+  small *p*-values can be printed in
+  formats like *p*<.001. Users can also
+  control the number of digits in the
+  printout. See the help page
+  of `print.summary.std_selected()`
+  to learn more about new arguments (0.2.8.9002).
+- Finalized to 0.2.9. (0.2.9)
+
 # stdmod 0.2.8
 
 - Added an R CMD check for noSuggests. (0.2.7.1)